+++ js/packages/eyes-api/src/input/RunnerOptions.ts
@@ -3,6 +3,9 @@ import * as utils from '@applitools/utils'
 export type RunnerOptions = {
   testConcurrency?: number
   fetchConcurrency?: number
+  /** @internal */
+  legacyConcurrency?: number
+  removeDuplicateTests?: boolean
 }
 
 /** @deprecated */