+++ js/packages/eyes-storybook/src/eyesStorybook.js
@@ -19,10 +19,10 @@ const {takeDomSnapshots} = require('@applitools/core');
 const {Driver} = require('@applitools/driver');
 const spec = require('@applitools/spec-driver-puppeteer');
 const {refineErrorMessage} = require('./errMessages');
-const {splitConfigsByBrowser} = require('./shouldRenderIE');
 const executeRenders = require('./executeRenders');
 const {makeCore} = require('@applitools/core');
 const {makeUFGClient} = require('@applitools/ufg-client');
+const makeGetStoriesWithConfig = require('./getStoriesWithConfig');
 
 const CONCURRENT_PAGES = 3;
 const MAX_RETRIES = 10;
@@ -65,19 +65,13 @@ async function eyesStorybook({
   if (config.puppeteerExtraHTTPHeaders) {
     await page.setExtraHTTPHeaders(config.puppeteerExtraHTTPHeaders);
   }
-  const core = await makeCore({
-    spec,
-    concurrency: testConcurrency,
-    logger,
-    agentId,
-  });
+  const core = await makeCore({spec, agentId, logger});
   const manager = await core.makeManager({
-    spec,
-    concurrency: testConcurrency,
-    agentId,
-    logger,
-    core,
     type: 'ufg',
+    settings: {
+      concurrency: testConcurrency,
+    },
+    logger,
   });
 
   const settings = {
@@ -87,6 +81,8 @@ async function eyesStorybook({
     agentId,
   };
   const [error, account] = await presult(core.getAccountInfo({settings, logger}));
+
+  const getStoriesWithConfig = makeGetStoriesWithConfig({config});
 
   if (error && error.message && error.message.includes('Unauthorized(401)')) {
     const failMsg = 'Incorrect API Key';
@@ -116,7 +112,7 @@ async function eyesStorybook({
   const pagePool = createPagePool({initPage, logger});
 
   const doTakeDomSnapshots = async ({page, renderers, layoutBreakpoints, waitBeforeCapture}) => {
-    const driver = await new Driver({spec, driver: page, logger}).init();
+    const driver = await new Driver({spec, driver: page, logger});
     const skipResources = client.getCachedResourceUrls();
     const result = await takeDomSnapshots({
       logger,
@@ -161,7 +157,18 @@ async function eyesStorybook({
       config,
     });
 
-    logger.log(`starting to run ${storiesIncludingVariations.length} stories`);
+    logger.log(
+      `there are ${storiesIncludingVariations.length} stories after filtering and adding variations `,
+    );
+
+    const storiesByBrowserWithConfig = getStoriesWithConfig({
+      stories: storiesIncludingVariations,
+      logger,
+    });
+
+    logger.log(
+      `starting to run ${storiesByBrowserWithConfig.stories.length} normal stories ("non fake IE") and ${storiesByBrowserWithConfig.storiesWithIE.length} "fake IE stories"`,
+    );
 
     const getStoryData = makeGetStoryData({
       logger,
@@ -197,14 +204,13 @@ async function eyesStorybook({
     });
 
     logger.log('finished creating functions');
-    const configs = config.fakeIE ? splitConfigsByBrowser(config) : [config];
+
     const [error, results] = await presult(
       executeRenders({
         renderStories,
         setRenderIE,
         setTransitioningIntoIE,
-        configs,
-        stories: storiesIncludingVariations,
+        storiesByBrowserWithConfig,
         pagePool,
         logger,
         timeItAsync,