+++ python/tests/coverage/selenium/conftest.py
@@ -2,11 +2,11 @@ import sys
 
 import selenium
 from pytest import fixture
+from selenium.common.exceptions import WebDriverException
 
 from applitools.selenium import BatchInfo, ClassicRunner, Eyes, StitchMode
 
 from .browsers import *
-from .devices import *
 from .sauce import pytest_collection_modifyitems, sauce_url
 
 batch_info = BatchInfo(
@@ -31,8 +31,8 @@ def driver_builder(chrome):
 
 @fixture
 def driver(driver_builder):
-    with driver_builder:
-        yield driver_builder
+    yield driver_builder
+    driver_builder.quit()
 
 
 @fixture