+++ python/tests/coverage/selenium/index.js
@@ -3,6 +3,10 @@ let dir = `https://raw.githubusercontent.com/applitools/sdk.coverage.tests/${ref
 module.exports = {
     name: "py-selenium",
     emitter: `${dir}/python/emitter.js`,
+    emitOnly: test => {
+      return !(test.features && test.features.includes("image")) &&
+          !("env" in test && "device" in test.env);
+    },
     overrides: [`${dir}/js/overrides.js`, `${dir}/python/overrides.js`],
     template: `${dir}/python/selenium-template.hbs`,
     tests: `${dir}/coverage-tests.js`,